requirements.txt
nicegui>=1.4.0
asyncpg>=0.29.0
anthropic>=0.7.0
replicate>=0.15.0
cloudinary>=1.36.0
pydantic>=2.4.0
python-multipart>=0.0.6
uvicorn>=0.23.0
python-dotenv>=1.0.0
psutil>=5.9.0